Renamed plugin to field type
