+++ python/src/aac/plugins/validators/field_type/_field_type.py
@@ -6,7 +6,7 @@ from aac.lang.language_context import LanguageContext
 from aac.plugins.validators import ValidatorFindings, ValidatorResult
 
 
-PLUGIN_NAME = "Subcomponents are models"
+PLUGIN_NAME = "Reference is type"
 
 
 def validate_subcomponent_types(
